Implemented VPTrack
Apart from implementing the (very few) methods of VPTrack, a bit of refactoring happened. voteOnTrack takes an integer instead of a bool. The client now keeps a currently playing track reference, necessary to tell whether a delete/vote operation is allowed or not, and a user name, for the same reason.