Implemented track-specific message sends. Turns out that this really needs a unique ID per track.
I previously identified the track by its position in the list, but that sucks. Votes can change the position in the list, and when to remove requests come at the same time, one of them may suddenly refer to something it wasn't mean to. So now a unique ID is used, which will probably simply be a single global counter in the server.